feat(Stripe): Add payment integration (#244)
## Overview:
- Add payment integration with Stripe when making purchase.
- What to test:

1. View paid ticket
2. Adjust quantity of paid ticket and observe change on summary page
3. Add credit card
4. Validate credit card, card must be correct format (card number, card
validate time, card CVV) . List cards for testing can be found
[here](https://stripe.com/docs/testing#cards)

**Note:** 
- Current payment API doesn't fully deployed yet so I assume it is
success and proceed.
- After BE deployed it will need to adjust to make flow work properly.

## Video:

https://github.com/lemonadesocial/lemonade-flutter/assets/108661349/858bb80c-9781-4c64-b420-84cabb09d716

---------

+++ lib/core/domain/event/input/redeem_tickets_input/redeem_tickets_input.dart
@@ -18,7 +18,7 @@ class RedeemTicketsInput with _$RedeemTicketsInput {
 @freezed
 class RedeemItem with _$RedeemItem {
   factory RedeemItem({
-    required double count,
+    required int count,
     @JsonKey(name: 'ticket_type') required String ticketType,
   }) = _RedeemItem;
